US04 - Ajout du footer avec liens de contact et réseaux sociaux, responsive
