+++ src/App.tsx
@@ -1,20 +1,20 @@
 import type React from "react";
 import Header from "./components/Header";
-import HexagonBubblesContainer from "./components/HexagonBubblesContainer";
+import HomePage from "./pages/HomePage";
 import LoginPage from "./pages/LoginPage";
 import "../src/App.css";
 import { Route, Routes } from "react-router-dom";
+import Footer from "./components/Footer";
 
 const App: React.FC = () => {
 	return (
 		<div className="app-container">
 			<Header />
-			<div className="d-flex justify-content-center align-items-center vh-100 bg-light">
-				<HexagonBubblesContainer />
-			</div>
 			<Routes>
+				<Route path="/" element={<HomePage />} />
 				<Route path="/login" element={<LoginPage />} />
 			</Routes>
+			<Footer />
 		</div>
 	);
 };
