+++ src/styles/Footer.css
@@ -1,0 +1,31 @@
+.footer {
+	background-color: #1e1e1e;
+	color: white;
+	padding: 1rem 0;
+	margin-top: auto;
+}
+
+.footer a {
+	color: #ffffff;
+	text-decoration: none;
+	transition: color 0.3s ease;
+	display: flex;
+	align-items: center;
+	gap: 8px;
+	justify-content: center;
+}
+
+.footer a:hover {
+	color: #0d6efd;
+}
+
+.footer svg {
+	font-size: 1.2rem;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+	.footer a {
+		font-size: 0.9rem;
+	}
+}
